remove duplicated mock files
