+++ tests/Company_Settings.spec.js
@@ -3,7 +3,7 @@ import { describe, it, expect, beforeEach, vi } from 'vitest'
 import { mount } from '@vue/test-utils'
 import { Suspense } from 'vue'
 import CompanySettings from '@/components/Company_Settings.vue'
-import { defaultGlobalStubs, createMockStore } from './test-utils.js'
+import { defaultGlobalStubs, createMockStore } from './helpers/test-utils.js'
 import { resolveAll } from './helpers/test-utils.js'
 
 // Mock dependencies at the top level
