+++ tests/ParcelStatuses_List.spec.js
@@ -3,7 +3,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { mount } from '@vue/test-utils'
 import { ref } from 'vue'
 import ParcelStatusesList from '@/components/ParcelStatuses_List.vue'
-import { defaultGlobalStubs } from './test-utils.js'
+import { defaultGlobalStubs } from './helpers/test-utils.js'
 
 // Mock functions at top level to avoid hoisting issues
 const getAllParcelStatuses = vi.hoisted(() => vi.fn())
@@ -342,3 +342,4 @@ describe('ParcelStatuses_List.vue', () => {
     })
   })
 })
+
