+++ tests/StopWords_List.spec.js
@@ -29,7 +29,7 @@ import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest'
 import { mount } from '@vue/test-utils'
 import { ref } from 'vue'
 import StopWordsList from '@/components/StopWords_List.vue'
-import { defaultGlobalStubs } from './test-utils.js'
+import { defaultGlobalStubs } from './helpers/test-utils.js'
 
 // Mock functions at top level to avoid hoisting issues
 const getAllStopWords = vi.hoisted(() => vi.fn())
@@ -515,3 +515,4 @@ describe('StopWords_List.vue', () => {
     })
   })
 })
+
