--- tests/__mocks__/localStorage.js
@@ -1,20 +1,0 @@
-import { vi } from 'vitest'
-
-// Create a localStorage mock that maintains state
-const createLocalStorageMock = () => {
-  let store = {}
-  return {
-    getItem: vi.fn((key) => store[key] || null),
-    setItem: vi.fn((key, value) => {
-      store[key] = value.toString()
-    }),
-    removeItem: vi.fn((key) => {
-      delete store[key]
-    }),
-    clear: vi.fn(() => {
-      store = {}
-    })
-  }
-}
-
-export default createLocalStorageMock
